Latest version.
git-svn-id: http://hdt-it.googlecode.com/svn/trunk@144 ca5865f0-ea67-0c57-7e18-94a3701c88e0

+++ hdt-lib/src/libdcs/VByte.cpp
@@ -67,7 +67,7 @@ namespace csd
 	}
 	
 	void 
-	VByte::bitset(uchar * e, size_t p) 
+	VByte::bitset(uchar * e, uint64_t p) 
 	{
         	e[p/8] |= (1<<(p%8));
     	}
